requirements.txt
discord.py
python-dotenv
flask
apscheduler
requests
beautifulsoup4
